delete http codes udpated

+++ jayfuncapp/products/api/src/functions/deleteProduct.ts
@@ -1,51 +1,61 @@
-import { CosmosClient, SqlQuerySpec } from '@azure/cosmos';
+import { CosmosClient, SqlQuerySpec } from '@azure/cosmos'
 import {
   app,
   HttpRequest,
   HttpResponseInit,
   InvocationContext,
-} from '@azure/functions';
-import Product from '../model/product';
+} from '@azure/functions'
+import Product from '../model/product'
 
 export async function deleteProduct(
   request: HttpRequest,
   context: InvocationContext
 ): Promise<HttpResponseInit> {
-  let deleted: any;
+  let deleted: any
   try {
-    context.log(`Http function processed request for url "${request.url}"`);
+    context.log(`Http function processed request for url "${request.url}"`)
 
-    const connectionString = process.env.CosmosDbConnectionString;
-    const databaseId = process.env.COSMOS_DATABASE_ID;
-    const containerId = process.env.COSMOS_CONTAINER_ID;
+    const connectionString = process.env.CosmosDbConnectionString
+    const databaseId = process.env.COSMOS_DATABASE_ID
+    const containerId = process.env.COSMOS_CONTAINER_ID
 
-    const client = new CosmosClient(connectionString);
-    const database = client.database(databaseId);
-    const container = database.container(containerId);
+    const client = new CosmosClient(connectionString)
+    const database = client.database(databaseId)
+    const container = database.container(containerId)
 
-    const id = request.query.get('id') || request.params.id;
-    context.log(`id : ${id}`);
+    const id = request.query.get('id') || request.params.id
+    context.log(`id : ${id}`)
 
     const querySpec: SqlQuerySpec = {
       query: 'SELECT * FROM c where c.id = @id',
       parameters: [{ name: '@id', value: id }],
-    };
-    const { resources } = await container.items.query(querySpec).fetchAll();
-    const product = resources[0] as Product;
-    context.log(`product id : ${product?.id}`);
-    context.log(`product categoryId : ${product?.categoryId}`);
-    deleted = await container.item(product?.id, product?.categoryId).delete();
+    }
+    const { resources } = await container.items.query(querySpec).fetchAll()
+    const product = resources[0] as Product
+    context.log(`product id : ${product?.id}`)
+    context.log(`product categoryId : ${product?.categoryId}`)
+    if (product?.id) {
+      deleted = await container.item(product?.id, product?.categoryId).delete()
+    } else {
+      return {
+        status: 204,
+        body: 'No product found for the given id and not deleted',
+        headers: {
+          'content-type': 'application/json',
+        },
+      }
+    }
 
-    context.log(`deleted : ${deleted}`);
+    context.log(`deleted : ${deleted}`)
   } catch (error) {
-    context.log(`Error: ${error}`);
+    context.log(`Error: ${error}`)
     return {
       status: 400,
       body: 'Failed to delete',
       headers: {
         'content-type': 'application/json',
       },
-    };
+    }
   }
   return {
     status: 200,
@@ -53,7 +63,7 @@ export async function deleteProduct(
     headers: {
       'content-type': 'application/json',
     },
-  };
+  }
 }
 
 app.http('deleteProduct', {
@@ -61,4 +71,4 @@ app.http('deleteProduct', {
   route: 'product/delete/{id}',
   authLevel: 'anonymous',
   handler: deleteProduct,
-});
+})
